Create tabbed initial page
This is currently very simple but the basic idea is that there is a page
to start a quiz (currently very empty) and another page to list all the
flags next to it. However these "pages" are not in the page stack but
they can be swiped back and forth just like in Sailfish's Settings app.
There isn't yet a nice header component for these.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>