qml: Change QuizTimer to use only one Timer
This fixes issue that there appears to be some time left in quiz timer
when it triggers. After this there is only one "reality" for quiz timer
and it shouldn't stop unless there is no time left.

Additionally make timer text become red once it runs out. Due to this
increase close timer to 1.5 seconds to make it appear more similar to
what it was before. Player's attention is likely drawn by the animating
text for the first 0.5 seconds instead of the list of choices.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>