qml: Move country list model to its own component
Move count list delegate model to its own component. Make loading of
that model asynchronous. Setting the model after it has loaded fixes a
crash on Xperia 10 II.

The crash was easiest to reproduce if when I set tab loader to only load
pages when their index is the current index. Then just switching to
countries tab would crash the app. Without this change it could be
reproduced with great effort by swiping to change the tab when the app
starts. That was not very reliable.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>