map: Improve colour scheme handling
Add invertedColour property to allow taking negative of the tiles. Also
handle overlay color change and minimap update properly.

Use timer to collapse texture creation calls. Reschedule it when more
properties change. This ensures we have the right values when the signal
is emitted and there are no duplicate calls to it that could mess up
things.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>