Add game modes
Default (and existing) behaviour is to save results without a name. In
that case "You" is credited as player. In party and shared device modes
a name is required. The difference between those is that party mode
discards any results when the application is closed. Anonymous mode
doesn't save any results.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>