assets: Add icons for covers
There is now new globe, map, flag and city icons that are meant to be
used as backgrounds for application cover. There are also medal, trophy
and big trophy icons for good results.

All icons are original works of mine. I based some of the designs on the
application icon to keep style consistent but that's also something I
created.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>