scoregraph: Fix rounding
In rounding up there was a corner case that some values weren't rounded
up. This happens because of truncation with integers. Convert to
floating point first.

In getDivision() there was another corner case that it was returning 1
when clearly the division points could be further apart. This happened
when span divided by scaleFactor was very small. To fix this convert to
floating point and check that we return at least scaleFactor.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>