scoregraph: Move squashing to background thread
This seems to be the most work in ScoreGraph but there are still other
heavy things happening there when there is a lot of data. However it's
quite difficult to start splitting anything out of updatePolish()
because those depend on members of ScoreGraph, or from paint() which
just has to happen there. Anyway I should profile the code before I
attempt to improve this further.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>