scoregraph: Fix corner cases with month texts
The text was too much on the right when the game is played on the very
first days of a month. Fix that by adjusting the limit on the right
side.

Always add the leftmost text if there is space. This ensures that there
is always at least one month text. Also adjust the logging a bit as
there can be different number of month lines and texts.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>