qml: Text input to QuizPage
This will be part of expert preset. If choices is set to zero (meaning
no choices but the player has to know), an input text box is shown
instead. Writing to that will filter the list of countries and show a
list of choices. The list is deliberately displaying only 1.5 choices so
that player cannot use it to guide their writing too much but can also
know that the list is scrollable.

Signed-off-by: Tomi Leppänen <tomi@tomin.site>